requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
attrs==21.2.0
    # via pytest
click==8.0.1
    # via code-annotations
code-annotations==1.1.2
    # via -r requirements/test.in
coverage==5.5
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-model-utils
    #   djangorestframework
    #   edx-django-utils
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-model-utils==4.1.1
    # via -r requirements/base.txt
django-waffle==2.2.1
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
djangorestframework==3.12.4
    # via -r requirements/base.txt
edx-django-utils==4.2.0
    # via -r requirements/base.txt
iniconfig==1.1.1
    # via pytest
jinja2==3.0.1
    # via code-annotations
markupsafe==2.0.1
    # via jinja2
newrelic==6.4.4.161
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==21.0
    # via pytest
pbr==5.6.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==0.13.1
    # via pytest
psutil==5.8.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==2.12.1
    # via -r requirements/test.in
pytest-django==4.4.0
    # via -r requirements/test.in
python-slugify==5.0.2
    # via code-annotations
pytz==2021.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==5.4.1
    # via code-annotations
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==3.3.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
